CORCI-436 ci: Initial pipeline support for IOF.
Add a Jenkinsfile to drive Jenkins
Update scons_local URL and reference.
Add Dockerfiles
Change testing so internals-tracing defaults to on, as this is now
run via nosetests where it was disabled.

This is a minimal working set of CI, where it runs checkpatch,
builds on CentOS and Ubuntu, and runs the single-node tests on
CentOS.

Not included are clang, and valgrind or fault injection tests
or the unit tests at this stage.

Change-Id: Ia2b2b05176f306382928ae338eebe150e5edb4be
Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>