IOF-343 test: Reenable nosetests for IOF testing.
Rename an internal function in the alloc_fail test so it's not
run as part of the nosetests.

nose auto-detects "tests" to run, by looking at function/object
names to know what to run, the recent change to alloc_fail
introduced a method which matched the regexp used.  Technically
this was correct, however the alloc_fail test is very different
and takes a long time to run, so disable it by renaming the function.

This means that once again the single-node tests can be run simply
by running:

$ nosetests3 --exe

Change-Id: I576a2959d14d41cc0e8809c46c058cca8704e740
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33600
Tested-by: Jenkins
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>