IOF-343 build: Remove check for d_log_sync_mask() function.
The prototype of this function has changed and configure code was added
to IOF to work with either the old or the new options, however this was
some time ago and IOF doesn't work with older versions of CaRT now so
remove the check and just code for the new version.

Remove the helper function from log.c as it only had two lines so it's
simpler to just perform it in-line.  Locks aren't really useful here
either so remove that as well.

Change-Id: I18750f062e11bbfa0621975a8a1ac1fbc673908e
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33607
Tested-by: Jenkins
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>