IOF-297 cart: Use proto_query() function to discover server version.
When registering the RPCs on the client use the protocol query code
to check supported versions on the server.

Only one version is implemented currently, so this is a simple check
rather than doing anything with the result but it does mean that
mismatch is detected and IOF will not start against incorrect server
versions.

Change-Id: I77a011b91dadf7b6777b931c4757b74372ccb0f1
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33630
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>
Reviewed-by: Dmitry Eremin <dmitry.eremin@intel.com>
Tested-by: Jenkins