IOF-343 CaRT: Split read/write into a new protocol and use boost macros
Split the read/write RPCs out from the core protocol into a new IO
protocol so they can be versioned together, and crucially not have
to change when the rest of the RPCs are changing.  As the interception
library now only registers these two RPCs it means it's much more
version agnostic when it comes to server versions.

Use the new CaRT macros for the read/write RPCs, and the embedded
xtvec struct.

Change-Id: I0c64f56fcda7f0157050db1b997d94763e0f6dcf
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33750
Tested-by: Jenkins
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>