CART-459 testing: Improve fault injection to test to abort on ignored faults.
Previously this test would check that all memory locations would be
freed after an allocation failure, but it was not checking that
allocation failures themselves were actually propagated.
After seeing a fault injected continue tracing the log file for
the success case, and if both are found then fail the test for not
propagating the error.

This catches some cases where return codes are not checked properly
which could lead to inconsistent behaviour.

Change-Id: Iea78af16ba5fa40242d04f0b13c086e92cb5b02d
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33886
Tested-by: Jenkins
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>