IOF-343 build: Update to latest scons_local
Removes deprecated preload function, replaced with load_definitions

Security checks don't like use of exec so just removing it

Also, moved ompi to end of prebuild line....this was a bug since
cart will build it and we want cart's version of it, not master.
I'd remove it entirely but there is still a bit of a bug as the
.build_vars script doesn't include the path if it's not in the list.
I'll leave that for another day.

Change-Id: Ifb650761f4d5282a9238744b944879dde70691e3
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33891
Tested-by: Jenkins
Reviewed-by: Ashley Pittman <ashley.m.pittman@intel.com>