IOF-343 testing: Fail tests which have errors in log files.
Extend the log checking code to fail tests which have error.  Make
this configurable and disable it for the CNSS logs for failover tests
which are expected to have errors, and whitelist the opc_lookup()
function which is known to log errors, and will be resolved when we
remove the lecacy features.

Change-Id: Ied45265ad044154ff84f983c2d3d5ffc96766e8b
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33947
Tested-by: Jenkins
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>