IOF-343 cnss: Change find_inode() function to operate on a request.
There is currently only one call-site for this function from
iof_fs_send() so change the args around to be cleaner, allowing
easier re-use from other parts of the code.

Remove the find_gah_ref() function entirely as it's not used, and clean
up find_gah_internal() now it doesn't need to support that use-case.

Change-Id: I8a29810ab4dd61a34a2a65d38308b1dcb6f96f56
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.hpdd.intel.com/34063
Tested-by: Jenkins
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>